requirements.txt
fastapi[all]==0.99.1
uvicorn[standard]==0.23.2
pdfplumber==0.11.9
pandas==2.2.3
openpyxl==3.1.2
python-multipart==0.0.6
aiofiles==23.1.0
Pillow==10.1.0
tqdm==4.66.0
opencv-python-headless==4.8.1.78
# PaddleOCR - lightweight OCR (prebuilt binaries)
paddleocr>=2.7.0
# Table detection
camelot-py[cv]>=0.10.1
# Fallback OCR
pytesseract>=0.3.10
# PDF image conversion
pdf2image>=1.16.0
